api/requirements.txt
fastapi
uvicorn
watchfiles
pypdf
python-multipart
asyncpg
spacy
httpx
PyJWT
bcrypt
